Epidemics

Documentation for Epidemics.

Benchmarking

This section shows some benchmarking.

julia> using BenchmarkTools
julia> using Epidemics
julia> # an epidemic of 500 days
       time_end = 500.0500.0
julia> # benchmark the default model with 3 age groups, intervention, and vaccination
       @benchmark epidemic_default(time_end=time_end, increment=1.0)BenchmarkTools.Trial: 488 samples with 1 evaluation.
 Range (min … max):   7.803 ms … 19.098 ms  ┊ GC (min … max): 0.00% … 31.33%
 Time  (median):      9.543 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   10.245 ms ±  1.974 ms  ┊ GC (mean ± σ):  5.35% ± 10.98%

          ▇▇█▅▂
  ▂▂▁▃▅▄▇██████▇▇▄▄▅▄▃▃▄▄▃▃▂▂▁▁▂▂▂▁▂▁▁▁▂▂▁▂▁▃▂▃▃▂▃▂▃▃▁▃▁▂▂▃▁▂ ▃
  7.8 ms          Histogram: frequency by time        16.9 ms <

 Memory estimate: 4.19 MiB, allocs estimate: 73844.

Epidemics.Npi

Epidemics.Population

Epidemics.Vaccination

Epidemics.default_contact_matrix

Epidemics.default_initial_conditions

Epidemics.epidemic_default

Epidemics.epidemic_default_ode!

Epidemics.infectious_period_to_gamma

Epidemics.preinfectious_period_to_alpha

Epidemics.prepare_args_default

Epidemics.prepare_contact_matrix

Epidemics.prepare_initial_conditions

Epidemics.r0_to_beta

Epidemics.Npi — Type

Npi(timebegin, timeend, contact_reduction)

A structure to hold the end points and strength of a non-pharmaceutical intervention.

Epidemics.Population — Type

Population(name, demographyvector, initialconditions, contact_matrix)

A structure to hold population characteristics, including:

'name': A name for the population.

'demography_vector': A numeric vector of the number of individuals in each

age or demographic group of the population.

'initial_conditions': A numeric matrix representing the proportions of each

age or demographic group that are in one of the epidemiological compartments.

'contact_matrix': A matrix giving the contacts between the demographic groups

in the population. Must be a square matrix.

Epidemics.Vaccination — Type

Vaccination(timebegin, timeend, ν)

A structure to hold the end points and strength of a non-pharmaceutical intervention.

Epidemics.default_contact_matrix — Method

default_contact_matrix(; n_groups=3)

Create a uniform contact matrix.

Named Arguments

n_groups: A Number for the number of demographic groups in the population.

Returns

A matrix of ones with the dimensions n_groups * n_groups.

Epidemics.default_initial_conditions — Method

default_initial_conditions(; n_groups=3, p_infected=[1e-6, 1e-6, 1e-6],
    p_exposed=[0.0, 0.0, 0.0])

Create some useful initial conditions for the default SEIRV epidemic model.

Named Arguments

n_groups: A Number for the number of demographic groups in the population.

p_infected: The proportion of each demographic group that is infected and also infectious. This is expected to be a Vector of Numbers with each value between 0.0 and 1.0.

p_exposed: The proportion of each demographic group that is exposed but not yet infectious. This is expected to be a Vector of Numbers with each value between 0.0 and 1.0.

Returns

A matrix with the dimensions n_groups * 5, with each row representing a demographic groups and each column representing one of the five epidemiological compartments of the default model.

Epidemics.epidemic_default — Method

epidemic_default(population, infection, intervention,
    time_end, increment
)

Model the progression of an epidemic, with age- or demographic-group specific contact patterns and proportions, non-pharmaceutical interventions with group- specific effects, and group-specific vaccination regimes.

Epidemics.epidemic_default_ode! — Method

epidemic_default_ode!(du, u, parameters, t)

A simple SEIRV epidemic model function that allows for multiple demographic groups. This function is intended to be called internally from epidemic_default.

The function expects the parameters argument to be a four element vector with the following elements:

a Population object with a prepared contact matrix, see [Population]

(@ref);

$\beta$, the transmission rate;

$\alpha$, the rate of conversion from exposed to infectious;

$\gamma$, the rate of recovery;

a matrix specifying the contacts between demographic groups;

an Npi object specifying the intervention applied to each age

group, see Npi;

a Vaccination object, see Vaccination;

Epidemics.infectious_period_to_gamma — Method

infectious_period_to_gamma(; infectious_period)

Calculate the recovery rate of the infection $\gamma$ from the infectious period.

Named Arguments

infectious_period: A single number for the mean duration in days that

individuals are infectious.

Returns

A single number representing the recovery rate of the infectious $\gamma$

Epidemics.preinfectious_period_to_alpha — Method

preinfectious_period_to_alpha(; r0, preinfectious_period)

Calculate the rate of transition from the 'exposed' to the 'infectious' compartment, $\alpha$ from the mean period between exposure and the initial occurrence of symptoms.

Named Arguments

preinfectious_period: A single number for the mean duration in days between

individuals being exposed to infection and becoming infectious.

Returns

A single number representing the transmission rate of the infectious $\alpha$

Epidemics.prepare_args_default — Method

prepare_args_default(parameters)

A function to prepare the parameters passed to the default epidemic model. This function is intended to be called internally from epidemic_default.

Named arguments

The function expects the parameters argument to be a four element tuple with the following elements:

population, for the population affected by the epidemic;

infection, for the infection causing the epidemic;

intervention, for any Npis applied to the epidemic;

vaccination, for any vaccinations applied to the population;

Return

Returns a tuple of the parameters for the default epidemic model.

Epidemics.prepare_contact_matrix — Method

prepare_contact_matrix(; population)

Prepare a population contact matrix for an epidemic model.

Named Arguments

population: A Population object with information about the population affected by the epidemic. Must include a contact matrix, and a demography vector.

Returns

A matrix of ones with the dimensions n_groups * n_groups.

Epidemics.prepare_initial_conditions — Method

prepare_initial_conditions(; population)

Prepare initial conditions from a Population object.

Named Arguments

population: A Population object with information about the population affected by the epidemic. Must include a matrix of initial conditions, and a demography vector.

Returns

A matrix with the initial conditions scaled by the demography vector.

Epidemics.r0_to_beta — Method

r0_to_beta(; r0, infectious_period)

Calculate the transmission rate $\beta$ from the basic reproductive number $R_0$ and the mean infectious period.

Named Arguments

r0: A single number for the basic reproduction number of an infection.

infectious_period: A single number for the mean duration in days that

individuals are infectious.

Returns

A single number representing the transmission rate of the infection $\beta$