Epidemics

Documentation for Epidemics.

Epidemics.Npi

Epidemics.Population

Epidemics.Vaccination

Epidemics.epidemic_default

Epidemics.epidemic_default_ode!

Epidemics.prepare_args_default

Epidemics.Npi — Type

Npi(timebegin, timeend, contact_reduction)

A structure to hold the end points and strength of a non-pharmaceutical intervention.

Epidemics.Population — Type

Population(name, demographyvector, initialconditions, contact_matrix)

A structure to hold population characteristics, including:

'name': A name for the population.

'demography_vector': A numeric vector of the number of individuals in each

age or demographic group of the population.

'initial_conditions': A numeric matrix representing the proportions of each

age or demographic group that are in one of the epidemiological compartments.

'contact_matrix': A matrix giving the contacts between the demographic groups

in the population. Must be a square matrix.

Epidemics.Vaccination — Type

Vaccination(timebegin, timeend, ν)

A structure to hold the end points and strength of a non-pharmaceutical intervention.

Epidemics.epidemic_default — Method

epidemic_default(population, infection, intervention,
    time_end, increment
)

Model the progression of an epidemic, with age- or demographic-group specific contact patterns and proportions, non-pharmaceutical interventions with group- specific effects, and group-specific vaccination regimes.

Epidemics.epidemic_default_ode! — Method

epidemic_default_ode!(du, u, parameters, t)

A simple SEIR epidemic model function that allows for multiple demographic groups. This function is intended to be called internally from epidemic. The function expects the parameters argument to be a four element vector with the following elements: - $\beta$, the transmission rate; - $\alpha$, the rate of conversion from exposed to infectious; - $\gamma$, the rate of recovery; - a matrix specifying the contacts between demographic groups; - an Npi object specifying the intervention applied to each age group, see Npi

Epidemics.prepare_args_default — Method

prepare_args_default(parameters)

A function to prepare the parameters passed to the default epidemic model. This function is intended to be called internally from epidemic. The function expects the parameters argument to be a four element tuple with the following elements: - population, for the population affected by the epidemic; - infection, for the infection causing the epidemic; - intervention, for any Npis applied to the epidemic; - vaccination, for any vaccinations applied to the population;