Epidemics

Documentation for Epidemics.

Epidemics.intervention

Epidemics.epidemic

Epidemics.seir!

Epidemics.intervention — Type

npi(timebegin, timeend, contact_reduction)

A structure to hold the end points and strength of a non-pharmaceutical intervention.

Epidemics.epidemic — Method

epidemic(model_name, init, contact_matrix, demography_vector, r0,
    preinfectious_period, infectious_period, interv, time_end, increment
)

Model the progression of an epidemic, with age- or demographic-group specific contact patterns and proportions, epidemiological parameters, and interventions.

Epidemics.seir! — Method

seir(du, u, parameters, t)

A simple SEIR epidemic model function that allows for multiple demographic groups. This function is intended to be called internally from epidemic. The function expects the parameters argument to be a four element vector with the following elements: a vector of beta, the transmission rate, where each element of the vector represents the transmission rate of the pathogen within a specific demographic group; a vector of beta2, the group-specific rate of conversion from exposed to infectious; a vector of gamma, the group-specific rate of recovery; a matrix specifying the contacts between demographic groups.